feat(crypto-paper): add repository root to PATH
This commit introduces an extensible mechanism to run script when PowerShell is launched from Visual Studio Code in a repository created from `crypto-paper`. The script currently adds the repository root to the `PATH` environment variable (with detection of various errors) and changes the location to the `paper` directory if the shell is started at the repository root.