fix(crypto-paper): fix bug in `opens-on` option
This commit fixes a bug in detecting the current page number. The counter `\c@page` is "delayed" in TeX in the sense that it does not always reflect the page number the current point will be on, due to how TeX ships pages out (and also elastic `\vskip`s). When `opens-on=right`, the logic is as follows now:
1. Force page number synchronization with `\clearpage` (this is OK since when `opens-on=right`, the title page and the table of contents are on pages separate from the main text).
2. Test whether the current page (now blank) is even-numbered, if so, advance to the next page.
3. Renumber the current page to 1 and allow main text to be typeset.

It also improves code style in `\laomian@tableofcontents@typeset` by first moving to the new page then switching off the hyperlink colors.