feat(crypto-paper): better hyper targets
This commit introduces `\HyperTargetEmbedRaised`, `\yesnumbertarget`, `\yesnumbertag`. It also updates the example usage of hyperlinked proofs.

The macro `\HyperTargetEmbedRaised{length}{name}` creates a raised link. It is intended to be used when `\HyperTargetToThisLine{name}` does not work. This known to happen when the link is inside either a display equation or a tabular cell in a figure, hence "embed".

The macros `\yesnumbertarget\label{eqn:name}` and `\yesnumbertag` should be used in starred display equations (notably `\begin{align*}`).
- The naming is inspired by `\nonumber` in non-starred display equations, which prevents the equation counter from incrementing and suppresses the equation tag (equation number in parentheses shown next to the line of equation). The macros achieve exactly the opposite effect and can be used to improve hyperlinking.
- The effect of `\yesnumbertarget\label{eqn:name}` is to increment the equation counter and place a hyper target *at its location*.
- The effect of `\yesnumbertag` is to show the equation tag with the current equation counter value.
- The reason to use this pair instead of the non-starred display equation environments is as follows.
  - In non-starred display equations, the hyper target is placed around the equation tag. This is problematic if the PDF is over-zoomed-in --- clicking on a hyperlinked reference will lead the user to the end of that line before the equation tag, not the start of the line (preferable).
  - Even if the PDF is not over-zoomed-in, e.g., when it is fit to window width, suppose we have a two-line equation in `\begin{align*}`, the first line has `\nonumber`, and the label is placed on the second line, then clicking on a hyperlinked reference will lead the user to the second line, not the first line (preferable).
  - By explicitly placing `\yesnumbertarget\label{eqn:name}`, the author gets to control the precise target so that it can be the beginning of the line of equation and the first line of a multi-line equation.
  - The macros can also be useful in case only a few lines should be numbered in a long series of aligned equations. Instead of too many `\nonumber`, we can issue a few `\yesnumbertarget`, `\yesnumbertag`.

Hyper targets of proofs should be the first item inside a `\begin{proof}`, instead of part of the proof title as done previously. At the beginning of a proof, TeX is in vertical mode and `\HyperTargetToThisLine` will put the target to the top-left of the current line. This is better than putting the target after "(" in "Proof (Theorem 1)." In addition, putting the target in front of "P" by making it part of the environment heading could cause other issues with `\label` inside a float inside the proof.