feat(dissertation-uw): improve footnote targets
This commit patches the footnote commands to put the hyper targets of footnotes at better locations. By default, they are located at the beginning of footnote text --- jumping to them does not reveal the footnote number or the beginning of the succeeding lines if the PDF is overly zoomed in. Now the targets are at the beginning of the first line of footnote text.

The standard `book` class makes the footnote text start at 1.8em on the first line. The `hyperref` package places the anchor (target) as the first thing in the footnote text. This patch dedents the footnote text including the anchor (target), then re-indents the actual text. The amount is 1.805em to take kerning into account so that the left edge does not cut into negatively kerned letters.

The version for normal footnotes (`\@footnotetext`) is tested on a working document to ensure it works and does not perturb typesetting results. The version for mini-page footnotes (`\@mpfootnotetext`) is not tested since it is not used in the working document.