fix!(dissertation-uw): spacing and layout
This commit brings the layout to compliance with UW Graduate School and ProQuest requirements.
- Everything except for page numbers is at least 1 inch away from the paper border.
- Page numbers are at least 0.75 inches away from the paper border.
It also changes how the type centers are vertically aligned (stable location of the text area).

This commit also adds space to list of figures and list of tables (consistent with subsection spacing in the table of contents).

BREAKING-CHANGE: The height of type center is made smaller. Manuscripts, especially large figures/tables, may have to be readjusted.